add userid

+++ model/index.js
@@ -1,18 +1,18 @@
 const Contact = require('./schemas/contact')
 
 
-const listContacts = async () => {
- const result = await Contact.find({})
+const listContacts = async (userId) => {
+ const result = await Contact.find({owner: userId})
  return result
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
+const getContactById = async (userId, contactId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId })
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findByIdAndRemove({ _id: contactId })
+const removeContact = async (userId, contactId) => {
+  const result = await Contact.findByIdAndRemove({ _id: contactId, owner: userId })
   return result
 }
 
@@ -22,18 +22,24 @@ const addContact = async (body) => {
 }
 
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    {
+       _id: contactId,
+       owner: userId,
+       },
     { ...body },
     { new: true }
   )
   return result
 }
 
-const updateContactStatus = async (contactId, body) => {
+const updateContactStatus = async (userId, contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    {
+       _id: contactId,
+      owner: userId,
+    },
     { ...body },
     { new: true }
   )
